tests/wpt/tests/intersection-observer/scroll-margin-dynamic.html

<!DOCTYPE html>
<meta name="viewport" content="width=device-width,initial-scale=1">
<link rel="help" href="https://www.w3.org/TR/intersection-observer/#dom-intersectionobserver-scrollmargin">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="./resources/intersection-observer-test-utils.js"></script>

<style>
#scroller { width: 100px; height: 100px; overflow: hidden; background-color: gray; }
#spacer { width: 50px; height: 120px; }
#target { width: 50px; height: 50px; background-color: green; }
</style>

<div id=scroller>
  <div id=spacer></div>
  <div id=target></div>
</div>

<script>
let entries = [];

window.onload = function() {
  runTestCycle(testNoInitialIntersection, "Test no initial scroll margin intersection");

  const observer = new IntersectionObserver(
    es => entries = entries.concat(es),
    {
      scrollMargin: "10px"
    }
  );
  observer.observe(target);
};

function testNoInitialIntersection() {
  assert_equals(entries.length, 1, "IntersectionObserverEntryCount");
  assert_false(entries[0].isIntersecting, "isIntersectingInitial");
  assert_approx_equals(entries[0].intersectionRatio, 0.0, 0.001, "intersectionRatioInitial");

  scroller.scrollTop = 20;
  runTestCycle(testIntersectionAfterScroll, "Test scroll margin intersection after scrolling");
}

function testIntersectionAfterScroll() {
  assert_equals(entries.length, 2, "IntersectionObserverEntryCount");
  assert_true(entries[1].isIntersecting, "isIntersecting");
  assert_approx_equals(entries[1].intersectionRatio, 0.2, 0.001, "intersectionRatio");
}
</script>
