LayoutTests/imported/w3c/web-platform-tests/intersection-observer/scroll-margin-dynamic-expected.txt


PASS Test no initial scroll margin intersection
FAIL Test scroll margin intersection after scrolling assert_approx_equals: intersectionRatio expected 0.2 +/- 0.001 but got 0
